+++ MarioProject/ItemAniIDs.h
@@ -1,9 +1,9 @@
 #pragma once
 //#include "AnimationIDs.h"
 
-constexpr int ID_ANI_SUPER_LEAF_ITEM		= 2e6;
-constexpr int ID_ANI_SUPER_LEAF_ITEM_LEFT	= 2e6;
-constexpr int ID_ANI_SUPER_LEAF_ITEM_RIGHT	= 2e6 + 100 ;
+constexpr int ID_ANI_SUPER_LEAF_ITEM		= static_cast<int>(2e6);
+constexpr int ID_ANI_SUPER_LEAF_ITEM_LEFT	= static_cast<int>(2e6);
+constexpr int ID_ANI_SUPER_LEAF_ITEM_RIGHT	= static_cast<int>(2e6 + 100);
 
-constexpr int ID_ANI_SUPER_MUSHROOM_ITEM	= 2e6 + 1e3 ;
-constexpr int ID_ANI_ONE_UP_MUSHROOM_ITEM	= 2e6 + 2e3 ;
+constexpr int ID_ANI_SUPER_MUSHROOM_ITEM	= static_cast<int>(2e6 + 1e3);
+constexpr int ID_ANI_ONE_UP_MUSHROOM_ITEM	= static_cast<int>(2e6 + 2e3);
